+++ src/onetick/py/oqd/conftest.py
@@ -6,14 +6,8 @@ from onetick.py.otq import otq  # noqa: F401
 
 
 @pytest.fixture(scope="function", autouse=True)
-def session():
-    locator = otp.RemoteTS(
-        otp.LoadBalancing(
-            "development-queryhost.preprod-solutions.parent.onetick.com:50015",
-            "development-queryhost-2.preprod-solutions.parent.onetick.com:50015"
-        )
-    )
-    cfg = otp.Config(locator=locator)
+def session(cloud_server):
+    cfg = otp.Config(locator=cloud_server)
 
     with otp.Session(cfg):
         yield